Updated adafruit_ds1307.py to allow Square Wave settings
The existing code relied on checking the values of the square wave register, and confirming that the default of 0x#3 was set.  These values are user selectable and should not be used to validate the device is an DS1307.
I followed the same logic, but applied it to the fields that are guaranteed to always be 0 in the DS1307 manual, instead of using user-changeable values.